Remove trips with only one stop from stop_times
To fix the situation, I think, arising from having one stop in the
geographical range of interest, but then all others outside of it. This
throws up lots of warnings in validation.

Now we throw these away as they're useless anyway.

So, what we should be keeping now are whole trips internally within a
selected region, and the parts of trips that go outside a selected
region. (The latter makes sense, they're still useful trips since we
could go from A to B in a region, but not be bothered about going to C
outside that region, if we want to consider all public transport options
within a geographic region.)